requirements.txt
tweepy==4.14.0
praw==7.7.1
supabase==2.3.0
python-dotenv==1.0.0
schedule==1.2.1
openai==1.12.0 